refactor(reticulum): drop ineffective peerStore dynamic imports
Extract pure path-slot helpers so pathMedium can import the store statically
without a cycle, silencing the Vite INEFFECTIVE_DYNAMIC_IMPORT warning.
